Do not accidentally overwrite an input file while renaming another.
If an input file would be renamed to the name of another input file,
rename it to something else until after the other input file has been
renamed.

For example, if we have 'a' and 'aa' as inputs where

    a => aa
    aa => aaa

we expect to have the files 'aa' and 'aaa' after running rename. Without
this special handling, 'a' would replace 'aa', then it would get renamed
again to 'aaa', losing the original 'aa' file.

Note that non-input files are (and should be) fair game for getting
overwritten.